[#5329] revert irods package due to namespace conflict
